Move JSONEpochConfigService to separate file
This was done to reduce the dependencies required to use it in a script
