+++ app/api/rewards/route.ts
@@ -1,7 +1,7 @@
 /**
  * @api {get} /api/rewards Get rewards for a given user and epoch based on their LP tokens
  */
-import {JSONEpochConfigService} from "@/src/models/campaigns/Campaign";
+import {JSONEpochConfigService} from "@/src/models/campaigns/JSONEpochConfigService";
 import {SentioJSONUserRewardsService} from "@/src/models/rewards/UserRewards";
 import {NotFoundError} from "@/src/utils/errors";
 import {NextRequest, NextResponse} from "next/server";
